add utility file for duplicate FAQ removal function
