+++ apps/webapp/scripts/sync-content.sh
@@ -311,6 +311,7 @@ import {
   unichainFaqItems
 } from './sharedFaqItems';
 import { getBundledTransactionsFaqItems } from './getBundledTransactionsFaqItems';
+import { deduplicateFaqItems } from './utils';
 
 export const getBalancesFaqItems = (chainId: number) => {
   const items = [
@@ -323,17 +324,7 @@ export const getBalancesFaqItems = (chainId: number) => {
     ...getBundledTransactionsFaqItems()
   ];
   
-  // Deduplicate by question (title), keeping the first occurrence
-  const seen = new Set<string>();
-  const deduplicatedItems = items.filter(item => {
-    if (seen.has(item.question)) {
-      return false;
-    }
-    seen.add(item.question);
-    return true;
-  });
-  
-  return deduplicatedItems.sort((a, b) => a.index - b.index);
+  return deduplicateFaqItems(items);
 };
 
 const generalFaqItems = ${match[1]};
@@ -373,6 +364,7 @@ import {
   optimismFaqItems,
   unichainFaqItems
 } from './sharedFaqItems';
+import { deduplicateFaqItems } from './utils';
 
 export const getSavingsFaqItems = (chainId: number) => {
   const items = [
@@ -385,17 +377,7 @@ export const getSavingsFaqItems = (chainId: number) => {
     ...(isL2ChainId(chainId) ? L2SavingsFaqItems : [])
   ];
   
-  // Deduplicate by question (title), keeping the first occurrence
-  const seen = new Set<string>();
-  const deduplicatedItems = items.filter(item => {
-    if (seen.has(item.question)) {
-      return false;
-    }
-    seen.add(item.question);
-    return true;
-  });
-  
-  return deduplicatedItems.sort((a, b) => a.index - b.index);
+  return deduplicateFaqItems(items);
 };
 
 const generalFaqItems = ${match[1]};
@@ -437,6 +419,7 @@ import {
   optimismFaqItems,
   unichainFaqItems
 } from './sharedFaqItems';
+import { deduplicateFaqItems } from './utils';
 
 export const getTradeFaqItems = (chainId: number) => {
   const items = [
@@ -449,17 +432,7 @@ export const getTradeFaqItems = (chainId: number) => {
     ...(isL2ChainId(chainId) ? L2TradeFaqItems : [])
   ];
   
-  // Deduplicate by question (title), keeping the first occurrence
-  const seen = new Set<string>();
-  const deduplicatedItems = items.filter(item => {
-    if (seen.has(item.question)) {
-      return false;
-    }
-    seen.add(item.question);
-    return true;
-  });
-  
-  return deduplicatedItems.sort((a, b) => a.index - b.index);
+  return deduplicateFaqItems(items);
 };
 
 const generalFaqItems = ${match[1]};
