+++ apps/webapp/src/data/faqs/getBalancesFaqItems.ts
@@ -13,6 +13,7 @@ import {
   unichainFaqItems
 } from './sharedFaqItems';
 import { getBundledTransactionsFaqItems } from './getBundledTransactionsFaqItems';
+import { deduplicateFaqItems } from './utils';
 
 export const getBalancesFaqItems = (chainId: number) => {
   const items = [
@@ -25,17 +26,7 @@ export const getBalancesFaqItems = (chainId: number) => {
     ...getBundledTransactionsFaqItems()
   ];
 
-  // Deduplicate by question (title), keeping the first occurrence
-  const seen = new Set<string>();
-  const deduplicatedItems = items.filter(item => {
-    if (seen.has(item.question)) {
-      return false;
-    }
-    seen.add(item.question);
-    return true;
-  });
-
-  return deduplicatedItems.sort((a, b) => a.index - b.index);
+  return deduplicateFaqItems(items);
 };
 
 const generalFaqItems = [
