+++ apps/webapp/src/data/faqs/getTradeFaqItems.ts
@@ -13,6 +13,7 @@ import {
   optimismFaqItems,
   unichainFaqItems
 } from './sharedFaqItems';
+import { deduplicateFaqItems } from './utils';
 
 export const getTradeFaqItems = (chainId: number) => {
   const items = [
@@ -25,17 +26,7 @@ export const getTradeFaqItems = (chainId: number) => {
     ...(isL2ChainId(chainId) ? L2TradeFaqItems : [])
   ];
 
-  // Deduplicate by question (title), keeping the first occurrence
-  const seen = new Set<string>();
-  const deduplicatedItems = items.filter(item => {
-    if (seen.has(item.question)) {
-      return false;
-    }
-    seen.add(item.question);
-    return true;
-  });
-
-  return deduplicatedItems.sort((a, b) => a.index - b.index);
+  return deduplicateFaqItems(items);
 };
 
 const generalFaqItems = [
